test(unit): refactor all test to match paths mappings
