+++ __tests__/unit/entities/Donor.spec.ts
@@ -1,6 +1,7 @@
 import { expect } from 'chai'
 
-import { User, Donor } from '@entities'
+import { User } from '@entities/User'
+import { Donor } from '@entities/Donor'
 
 describe('Donor Entity', () => {
   it('Should be able to create a new Donor and be instance from Entity', () => {
