+++ __tests__/unit/entities/Entity.spec.ts
@@ -1,6 +1,6 @@
 import { expect } from 'chai'
 
-import { BaseEntity } from '@entities'
+import { BaseEntity } from '@entities/BaseEntity'
 
 describe('Abstract Entity', () => {
   it('Should be able to create a new Entity without the values of id, created_at, updated_at', () => {
