+++ __tests__/unit/entities/Incident.spec.ts
@@ -1,6 +1,7 @@
 import { expect } from 'chai'
 
-import { BaseEntity, Incident } from '@entities'
+import { BaseEntity } from '@entities/BaseEntity'
+import { Incident } from '@entities/Incident'
 
 describe('Incident Entity', () => {
   it('Should be able to create a new Incident without the values of id, created_at, updated_at', () => {
