+++ __tests__/unit/entities/Ong.spec.ts
@@ -1,6 +1,7 @@
 import { expect } from 'chai'
 
-import { User, Ong } from '@entities'
+import { User } from '@entities/User'
+import { Ong } from '@entities/Ong'
 
 describe('ONG Entity', () => {
   it('Should be able to create a new ONG and be instance from Entity', () => {
